fix: calendar .is-today.is-focused should have blue focus ring, fixes #1051 (#1052)
See https://opensource.adobe.com/spectrum-css/calendar.html#focused

The grey background color for the .is-focused style on today's date is insufficient to communicate that the date has focus. We should use either the blue focus ring as we do for other focused styles, or to continue communicating that the date is today, some sort of double border.