fix: guard against operations on `null` collections
This resolves an issue where the `setFocus` method was attempting to add or remove classes from an element that doesn't always exist in our examples.

With the introduction of the `readonly` version of the Datepicker component, the `readonly` state does not implement the Picker Button, and therefore the code was attempting to add/remove classes from a `null` element.