fix: use stricter comparision for peerDependency version semver checks
This changes the way that we check to see if a package's `peerDependencies` match (or would be 'satisfied') by the version required in the package's `devDependencies`. Our previous comparison was not properly handling prerelease version satisfaction (or, at least, it wasn't handling it in a way that we needed it to). This change updates the `satisfies` args so that they compare exact version numbers (instead of the "^" range).