fix(illustratedmessage): workaround for safari currentcolor bug
Workaround for issue where Safari browser v16.3 will not change the
color of the fill/stroke that is set to currentcolor. It computes the
color of the parent selector setting the fill/stroke properties, rather
than the color of the current element.

This browser bug was fixed in the recent release of Safari 16.4.