fix(icon): correct storybook icon sizing substitution
Fix issue where use of Icon in other stories would not change its
sizing (scale) based on the value of the size arg. The conditional with
uiIcons.includes was comparing against the icon name which still
contained the sizing number, so was always failing.

Also adds some docs around this.