feat(table): high contrast mode styles
Styles for forced-colors mode (Windows High Contrast Mode). Includes use
of SelectedItem and SelectedItemText for showing the selected and
dropzone rows in blue. The forced-color-adjust: none was necessary to
add, otherwise the user agent was adding its own extra background color
to cell contents in addition to the cell background color.

Also reorganizes the passthrough mod for thumbnail.