fix(table): top alignment of checkboxes
Align checkboxes top the top of the cell, so they are in alignment when
there is wrapping text in other cells. Previously vertical alignment was
mismatched with middle and vertical. Simplifies some of this overall
vertical alignment and checkbox spacing.

Also makes sure corners are rounded for row dropzone used on the first
and last rows.