fix(search): simplify forced-colors mode fix to help with swc merge
Changes one of the forced-colors (Windows high contrast) mode fixes
for the overlapping close button, that was causing an issue on the SWC
draft merge VRTs. This accomplishes the fix is a different way, which is
simpler and does not involve any positioning changes. It now uses the
forced-color-adjust none to keep the transparent background on the close
button in high contrast mode.