feat(colorslider): reversed gradient for rtl support and docs update
Use same technique as SWC for reversing the gradient when using a RTL
base direction. The handles were flipping to the other side using the
new logical properties but the gradient was not.

Also updates documentation around this and improves overall docs
explanations. Changes first usage docs bullet point specifying
translateX, as this does not appear to be correct (SWC changes the
inset-inline-start and inset-block-end of the position absolute handle).

Fixes capitalization of component name in a few places to match the
sentence case specified in the Spectrum capitalization guidelines.