feat(table): finalize vertical alignment spacing with new tokens
Adjustments to make sure that cells maintain the same padding-block and
vertical alignment within the thumbnail variant. Cells without a
thumbnail in a row with thumbnails needed to align. Also makes
adjustments (a max() calc) around the fact that the disclosure icon can
sometimes be larger than the spacing defined for the thumbnail variant
density + sizing.