docs(menu): remove submenu from drill-in example
The displayed submenu for the drill-in example was not how submenus
should be displayed per the guidelines. They need to be positioned, and
are shown as being within another popover. Showing this adjacent menu
like this could cause some confusion as to its usage; the adjacent menu
was not positioned properly and does not have any separate styles within
the CSS for doing so.