refactor(table): remove some unnecessary nesting of style rules
Move some style rules outside of .spectrum-Table-body that do not need
that extra specificity since the .spectrum-Table-row is always within
the body. To help facilitate rules translating properly to SWC.

The :not rule was also not necessary for quiet, as the quiet border
radius custom property is already set to zero for quiet (and we want to
allow this to be moddable).