refactor(table): remove mask solution and add border around header
The solution to fake rounded corners and cover up table body scrolling
underneath was close to viable in Chrome and Firefox, but did not work
in Safari due to a Safari bug. There was also some minor fuzzyness
around corner pixels, which was more noticable in focus and dropzone
states.