refactor(table): change how scrollable wrapper is structured
Change markup for scrollable wrapper, so custom properties do not need
to be copied into it. Use root class and its variant classes on the main
wrapper, with an added Spectrum-Table-main for the main table.

Adjusts default scroll-padding-top with calculated value.

Changes tabindex behavior in Storybook to match docs examples.