feat(table): scrollable wrapper with sticky header
Solution to allow for a scrollable table body while keeping alignment
of table regardless of the length of cell content. Added because current
flex or grid solutions can result in misaligned cells when longer
content is displayed in a cell; especially at smaller screen sizes or
when using more columns. The only current viable solutions involve
manual matching adjustment of widths via JavaScript or with subgrid
(which is not yet ready for production).

This solution:
1. Adds a wrapper element with scrolling overflow that can have a fixed
   height set.
2. Sets the table header to be position sticky.
3. Adds a border around the header and sets a background color on it.

This update also includes:

feat(table): tabindex update for multi-select examples

Allow focus on checkbox instead of both the checkbox and the row, in
examples and Storybook. Matches with behavior in SWC component. Removes
tabindex from rows in multi-select variants with the checkbox cells.