docs(actionbar): update table example markup
Update Table example for Action bar. Action bar should not be nested
within the table markup. It throws off table display, is questionable
from a a11y perspective, and has been confirmed with design that it
should not appear on top.

Also adds missing Guidelines link and updates minimum version of table
to the current beta.