fix(stepper): adjustments for width differences
Using component-height-200 instead of component-height-100 was making
the component too much wider than the original. Now uses the stepper
height token, which is using component-height-100 at medium, and
adjustments for t-shirt sizing. This makes the larger sizes like XL
scale more appropriately in their width as well. Previously the
characters entered in the field would turn to ellipsis after 3 chars at
XL size.

Note: Medium version of component -should- be about 6px wider due to
wider in-field buttons, defined in tokens. Original medium was 72px.

Adjusts usage of some of the width tokens so that usage of
--mod-stepper-width and --mod-stepper-border-width will have the desired
effect.

Also needed to add border width to calc, so that 4 chars appear before
ellipsis at medium size, instead of 3. Was a matter of 1 or 2 pixels.