fix: exclude passthrough mods from generated mods.md
Components that set a subcomponent's mod property were showing them as
mods in the mods.md file. This adjusts the regex of our 'yarn docs:mod'
command, mod-extractor node task, so that it excludes found mod names
that are followed by a colon.

This includes running that command and the resulting modifications to
the various mods.md files.