refactor(treeview): fix heights and sizing of treeview items
Treeview items and the item link were previously not displaying at the
correct component-height values, especially at different t-shirt sizes.
Size was also not being passed down fully in Storybook, which was
masking the entire issue. The chevron indicator when properly sized was
not aligned properly on the left side of the item.

This fix should ensure the proper height for both elements at s, m, l,
and xl, and the aligned square chevron indicator. It also simplifies
and clarifies some of the negative margins and offsets that were being
used to push over the chevron.