refactor(treeview): remove negative block padding and fixes
Simplify block spacing to remove negative margin and negative offset.
Previously the indicator was displayed within the line-height, and had
negative spacings applied to re-align it and prevent it from making the
link taller, which was a bit overcomplicated and brittle.

Thumbnail:
Makes some adjustments to align this with the design. Changes the
thumbnail example to layer variant with is-selected to match with the
design file. This previously was showing the focused state for selected,
which has different visuals.

Standalone/Detached:
Standalone corner radius styles were not being applied and are now
fixed. Standalone was renamed to "detached" per the beta guidelines for
Thumbnail, so thatclass has been renamed and noted in the migration
guide.

Also simplifies styles for both the thumbnail and detached variants, to
just be changes in custom property values.