docs(checkbox): fix duplicate ids appearing in storybook
The same value was being used for the ID attribute on the checkbox
label and checkbox itself. This update makes sure they are both given
different IDs based on the passed in ID, by adding a suffix to the ID
used on the checkbox input.

This also fixes the markup previously showing an empty ID attribute on
the checkbox when the ID was not set.