chore: support .env template at root (#2255)
A utility that copies the .env settings from a user's root to the project to ease the following workflow:

git clean -dfx
yarn install

Checking out a clean repo is a critical part of the workflow but in doing so, it removes any uncommitted utilities such as the .env file with the Chromatic project token. Any time a yarn install is run, this utility script tries to copy variables listed in the example file from root.