docs(fieldgroup): remove form class not used for alignment
Remove the .spectrum-Form class from the Field group examples, as this
currently is not being used for alignment or with the correct markup
for Form (e.g. there is no .spectrumForm-item). It is not needed here
to demonstrate the Field group markup and CSS, which has its own flex
alignment used.

Previously the only thing this class was doing on this page was changing
the top and bottom margins of the form, while also oddly having a
display:flex item as a child of a display:table element. Form has its
own example page.