refactor(dropzone): update values for pointer-events custom prop
This removes the `--mod-*` prefixed custom properties for pointer-events used
within Drop Zone-nested Illustrated Messages. These custom properties were
preventing interactions within the Illustrated Message component.

Instead, over in Illustrated Message, we'll keep the `--mod-*` custom props
available in the stack defined for `pointer-events`, but we'll set a fallback
value of `auto`, which will preserve the ability for consumers to use the mod
but also get the default style if it is not set.