fix: mod-extractor support for multiple css files in root
Fixes the fact that if there are multiple CSS files in the root, such
as with the icon component (and commons), only the mods from the last
file found will be included. During migration changes to Icon, some mods
were being left out of the metadata when the utility was run.

The utility now bundles all parsed mods per component directory into an
array before exporting them. It also splits the utility into two
functions as it was quite large and the functionality was very
independent.