fix(picker): whcm - fix focus indicator and border colors
High contrast mode fixes for Picker. By default, the focus indicator
outline was showing on the picker when it did not have focus. The colors
were also reversed; these now mirror the focus indicator color and
border color combination shown in textfield when keyboard focused. Also
fixes some other high contrast colors that were a little off on hover
and when the picker was open.