feat(fieldlabel): form - replace table layout with grid (#2269)
* feat(fieldlabel): form - use grid instead of table

Refactor to modernize .spectrum-Form to use CSS grid instead of display
table, which helps get rid of the need for negative margins and
different ways of handling spacing between the two variants.

MOD PROPERTY CHANGES
Deprecates some Form --mod properties:
--mod-tableform-border-spacing
--mod-tableform-margin
--mod-tableform-margin-labels-above

The --mod-tableform-border-spacing mod has been renamed to
--mod-form-item-block-spacing

The margin mods are no longer necessary and removing them is in line
with other components which do not have mods for their root element
margins when not necessary.

Additional mods have been added for modifying the size of the the grid
tracks.

DOCS CHANGES
Removed usage of "placeholder" in form examples, as placeholder is no
longer recommended in guidelines or demoed for Textfield examples.

* docs(fieldgroup): remove form class not used for alignment

Remove the .spectrum-Form class from the Field group examples, as this
currently is not being used for alignment or with the correct markup
for Form (e.g. there is no .spectrumForm-item). It is not needed here
to demonstrate the Field group markup and CSS, which has its own flex
alignment used.

Previously the only thing this class was doing on this page was changing
the top and bottom margins of the form, while also oddly having a
display:flex item as a child of a display:table element. Form has its
own example page.

* docs(fieldlabel): form - example markup cleanup

- Use updated component markup for Stepper
- Use updated component markup for Picker
- Replace duplicate IDs with unique IDs

* docs(fieldlabel): form - storybook - update associated id for field

Update associated ID to work with separate PR #2271 made for textfield
and stepper to allow IDs on the inputs.

* docs(fieldlabel): form - labels for country

Include labels for country picker, connected to the ID of the button
within the picker.