fix(picker): key focus styles should not be used for focus class
The .is-focused class was being used as if it was .is-keyboardFocused,
in combination with :focus-visible in selectors. This class has been
removed as it is handled by :focus-visible on the button.

This fixes an existing issue on the docs site, where if you click the
picker twice, the focus indicator ring will appear. This is because of
the docs site adding the .is-focused class.

Updated the existing story that was labeled "Focused" that was showing
keyboard focused to use the play function, for Chromatic coverage
ref: https://www.chromatic.com/docs/hoverfocus/