refactor: run mods-extractor per component and better logging
Loop through assets by directory instead of all at once, so that we can
run this on a single package eventually as part of the build workflow.

Update console logging to give more info on what's going on.

Added exceptions for commons and overlay which don't have a
dist/index.css